yolo_service/requirements.txt
boto3==1.34.112
botocore==1.34.112
Flask==3.0.0
Flask-Cors==4.0.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.4
pickle4==0.0.1
psutil==5.9.7
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.5.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.0.229
uvicorn==0.27.1
